feat: 정적 메타데이터 생성
